# openapi/connector_mgmt-private.yaml
openapi: 3.0.0 # need this as first line to allow some IDEs to know this is an openapi document.
...
info:
  title: Connector Service Fleet Manager Private APIs
  version: 0.0.3
  description: Connector Service Fleet Manager apis that are used by internal services.
servers:
  - url: https://api.openshift.com
    description: Main (production) server
  - url: https://api.stage.openshift.com
    description: Staging server
  - url: http://localhost:8000
    description: localhost
  - url: /
    description: current domain
...
paths:
...
  '/api/connector_mgmt/v1/kafka_connector_clusters/{connector_cluster_id}/deployments/{deployment_id}/status':
    parameters:
      - name: connector_cluster_id
        description: The id of the connector cluster
        schema:
          type: string
        in: path
        required: true
      - name: deployment_id
        description: The id of the deployment
        schema:
          type: string
        in: path
        required: true
    put:
      tags:
        - Connector Clusters Agent
      operationId: updateConnectorDeploymentStatus
      summary: update the connector deployment status
      description: update the connector deployment status
      security:
        - Bearer: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectorDeploymentStatus'
        required: true
      responses:
        '200':
          description: Cluster status is updated
        '400':
          content:
            application/json:
              schema:
                $ref: 'kas-fleet-manager.yaml#/components/schemas/Error'
              examples:
                400InvalidIdExample:
                  $ref: '#/components/examples/400InvalidIdExample'
          description: id value is not valid
        '404':
          content:
            application/json:
              schema:
                $ref: 'kas-fleet-manager.yaml#/components/schemas/Error'
              examples:
                404Example:
                  $ref: 'kas-fleet-manager.yaml#/components/examples/404Example'
          # This is deliberate to hide the endpoints for unauthorised users
          description: Auth token is not valid.
        '410':
          description: deployment has been deleted
          content:
            application/json:
              schema:
                $ref: 'kas-fleet-manager.yaml#/components/schemas/Error'
components:
  schemas:
    MetaV1Condition:
      type: object
      properties:
        type:
          type: string
        reason:
          type: string
        message:
          type: string
        status:
          type: string
        lastTransitionTime:
          type: string
          deprecated: true
        last_transition_time:
          type: string
          #format: date-time # enable this?? or just set it as pure plain string?
...
    ConnectorDeploymentStatus:
      description: The status of connector deployment
      type: object
      properties:
        phase:
          type: string
        resource_version:
          type: integer
          format: int64
        available_upgrades:
          type: string
          deprecated: true
        operators:
          type: object
          properties:
            assigned:
              $ref: '#/components/schemas/ConnectorOperator'
            available:
              $ref: '#/components/schemas/ConnectorOperator'
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/MetaV1Condition'
...
    ConnectorOperator:
      description: identifies an operator that runs on the fleet shards used to manage connectors.
      properties:
        id:
          description: the id of the operator
          type: string
        type:
          description: the type of the operator
          type: string
        version:
          description: the version of the operator
          type: string
...
  securitySchemes:
    Bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  examples:
    400InvalidIdExample:
      value:
        id: "21"
        kind: "Error"
        href: "/api/kafkas_mgmt/v1/errors/21"
        code: "KAFKAS-MGMT-21"
        reason: "Bad request"
        operation_id: "1lWDGuybIrEnxrAem724gqkkiDv"